Allowing YTArray and YTQuantity to accept a dtype.
--HG--
branch : yt-3.0

+++ yt/utilities/orientation.py
@@ -45,9 +45,9 @@ class Orientation:
 
         # Make sure vectors are unitless
         if north_vector is not None:
-            north_vector = YTArray(north_vector,"")
+            north_vector = YTArray(north_vector, "", dtype='float64')
         if normal_vector is not None:
-            normal_vector = YTArray(normal_vector,"")
+            normal_vector = YTArray(normal_vector, "", dtype='float64')
 
         self.steady_north = steady_north
         if not np.dot(normal_vector, normal_vector) > 0:
